backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2
pytest==7.4.3
gunicorn==21.2.0
